Add ast nodes for various initial statements
