+++ src/main/kotlin/com/github/derg/transpiler/ast/Nodes.kt
@@ -12,21 +12,22 @@ package com.github.derg.transpiler.ast
 sealed class Node
 
 /**
- * While typically structures in the source code may be represented by a single node, sequences of nodes are used to
- * describe larger components such as modules, namespaces, scopes, and so on.
+ * Expressions are computable bits of code which resolves down to a single value and type. These code elements cannot be
+ * re-assigned to other values, and do not occupy any space in memory. Intermediary computations may be stored on the
+ * stack, although the final value will either be used as a parameter for a procedure call, or stored in a variable.
  */
-typealias Sequence = List<Node>
+sealed class Expression : Node()
+
+/**
+ * Statements are executable bits of code, which either performs an operation with a side effect, or determines the
+ * control flow. Examples include assigning a value to a variable and returning from a sub-routine, respectively. Note
+ * that expressions are not statements.
+ */
+sealed class Statement : Node()
 
 /**
  * Structural components within the source code must be defined to describe their behavior fully. While certain objects
  * such as functions may be declared before being defined, all such objects must be defined before being used. The
  * definition of the object varies from object to object.
  */
-sealed class Definition : Node()
-
-/**
- * Expressions are computable bits of code which resolves down to a single value and type. These code elements cannot be
- * re-assigned to other values, and do not occupy any space in memory. Intermediary computations may be stored on the
- * stack, although the final value will either be used as a parameter for a procedure call, or stored in a variable.
- */
-sealed class Expression : Node()
+sealed class Definition : Statement()
